Fix: rename BookRow to indicate it should be used with StockTable
