+++ apps/web-client/src/lib/components/Tables/InventoryTables/StockBookRow.svelte
