+++ apps/web-client/src/lib/components/Tables/InventoryTables/index.ts
@@ -1,4 +1,4 @@
-export { default as BookRow } from "./BookRow.svelte";
+export { default as StockBookRow } from "./StockBookRow.svelte";
 
 export { default as StockTable } from "./StockTable.svelte";
 export { default as InboundTable } from "./InboundTable.svelte";
